Refactor the code, getting rid of blueprints
Blueprints just weren't made for this use case.

They do register static files separately, but the reason they tend to be in
a separate directory is that they can be served with a dumb server. The way
Blueprints handle this does not help that; they might as well be served
by Flask itself. This doesn't hurt when building a frozen site, and Flask
exposes the helper function that serves them with the correct MIME type.

Using static files from templates also didn't work out well: it should be used
as url_for('blueprintname.static', ...). This can be shortened to
url_for('.static', ...) to get static files from the current blueprint.
But our existing code didn't serve course pages from within the blueprint:
the route was application-specific, forcing templates to either use the long
form, or a special helper, or even a raw relative URL. Most of the time,
it used a raw URL – which would almost certainly break whenever we'd switch
to dynamic pages and use a dumb server to serve static files from a separate
directory.

All in all, blueprints weren't made for this, and I'd rather spend time bending
Flask/Werkzeug to my will than bending an extra layer of magic that appears to
solve one of the things we want, but complicates others.