Fix Flask app name
Set Flask app name.to 'naucse'
This makes app.root_path the directory of the 'naucse' package,
rather than the current working direstory as before.
Move 'static' to its usual place in a Flask app.
Remove customization for the template loader, as the directory
is at the default place now.
Adjust pathnames to match.