Don't show session in breadcrumbs of "extra" lessons
Lessons that aren't in the manifest don't belong to any
session, but they are useful. (And to make rendering from
multiple repositories possible in the future, they need to
be in a course's namespace, so we can't just use "/lesson"
(the current canonical URL) for them.)

Don't show the session in the breadcrumbs for these.

Fixes: https://github.com/pyvec/naucse.python.cz/issues/247