converters: Pass contexts to all load/dump operations
The API version determines how data is loaded/dumped, so it
needs to be available or all the load/dump methods.
Currently the naucse API version is two numbers, expected to
increase linearly, but I could imagine e.g. branching versioning
in the future (for example with "preview feature" tags, like
what GitHub does in its API). That's why I don't use "version"
directly, but wrap it in an abstract "context".

The top-level dump() function now requires a version, and creates
a context.

The load() methods should not be called top-level, but always
via the top-level load() function (which reads the API version from
passed-in data, and creates a context).

Adjust models and tests.